Modify pinMapping to also include profiles
