+++ www/src/Services/WebApi.js
@@ -76,40 +76,40 @@ export const baseProfileOptions = {
 };
 
 export const basePeripheralMapping = {
-    peripheral: {
-        i2c0: {
-            enabled: 0,
-            sda: -1,
-            scl: -1,
-            speed: 400000,
-        },
-        i2c1: {
-            enabled: 0,
-            sda: -1,
-            scl: -1,
-            speed: 400000,
-        },
-        spi0: {
-            enabled: 0,
-            rx:  -1,
-            cs:  -1,
-            sck: -1,
-            tx:  -1,
-        },
-        spi1: {
-            enabled: 0,
-            rx:  -1,
-            cs:  -1,
-            sck: -1,
-            tx:  -1,
-        },
-        usb0: {
-            enabled: 0,
-            dp:  -1,
-            enable5v: -1,
-            order: 0,
-        }
-    }
+	peripheral: {
+		i2c0: {
+			enabled: 0,
+			sda: -1,
+			scl: -1,
+			speed: 400000,
+		},
+		i2c1: {
+			enabled: 0,
+			sda: -1,
+			scl: -1,
+			speed: 400000,
+		},
+		spi0: {
+			enabled: 0,
+			rx: -1,
+			cs: -1,
+			sck: -1,
+			tx: -1,
+		},
+		spi1: {
+			enabled: 0,
+			rx: -1,
+			cs: -1,
+			sck: -1,
+			tx: -1,
+		},
+		usb0: {
+			enabled: 0,
+			dp: -1,
+			enable5v: -1,
+			order: 0,
+		},
+	},
 };
 
 export const baseWiiControls = {
@@ -410,49 +410,6 @@ async function setPinMappings(mappings) {
 		});
 }
 
-async function getProfileOptions(setLoading) {
-	setLoading(true);
-
-	try {
-		const response = await axios.get(`${baseUrl}/api/getProfileOptions`);
-		let profileOptions = { ...baseProfileOptions };
-		response.data['alternativePinMappings'].forEach((altButtons, index) => {
-			for (let prop of Object.keys(altButtons))
-				profileOptions['alternativePinMappings'][index][prop].pin = parseInt(
-					response.data['alternativePinMappings'][index][prop],
-				);
-		});
-		setLoading(false);
-		return profileOptions;
-	} catch (error) {
-		console.error(error);
-		return false;
-	}
-}
-
-async function setProfileOptions(options) {
-	let data = {};
-	data['alternativePinMappings'] = [];
-	options['alternativePinMappings'].forEach((altButtons, index) => {
-		let altMapping = {};
-		Object.keys(options['alternativePinMappings'][index]).map(
-			(button, i) => (altMapping[button] = altButtons[button].pin),
-		);
-		data['alternativePinMappings'].push(altMapping);
-	});
-
-	return axios
-		.post(`${baseUrl}/api/setProfileOptions`, sanitizeRequest(data))
-		.then((response) => {
-			console.log(response.data);
-			return true;
-		})
-		.catch((err) => {
-			console.error(err);
-			return false;
-		});
-}
-
 async function getKeyMappings(setLoading) {
 	setLoading(true);
 
@@ -573,7 +530,7 @@ async function setPS4Options(options) {
 async function getWiiControls(setLoading) {
 	setLoading(true);
 
-    try {
+	try {
 		const response = await axios.get(`${baseUrl}/api/getWiiControls`);
 		setLoading(false);
 
@@ -586,7 +543,7 @@ async function getWiiControls(setLoading) {
 }
 
 async function setWiiControls(mappings) {
-    console.dir(mappings);
+	console.dir(mappings);
 
 	return axios
 		.post(`${baseUrl}/api/setWiiControls`, sanitizeRequest(mappings))
@@ -601,8 +558,8 @@ async function setWiiControls(mappings) {
 }
 
 async function getPeripheralOptions(setLoading) {
-    setLoading(true);
-    try {
+	setLoading(true);
+	try {
 		const response = await axios.get(`${baseUrl}/api/getPeripheralOptions`);
 		setLoading(false);
 
@@ -615,7 +572,7 @@ async function getPeripheralOptions(setLoading) {
 }
 
 async function setPeripheralOptions(mappings) {
-    console.dir(mappings);
+	console.dir(mappings);
 
 	return axios
 		.post(`${baseUrl}/api/setPeripheralOptions`, sanitizeRequest(mappings))
@@ -716,8 +673,6 @@ const WebApi = {
 	getCustomTheme,
 	setCustomTheme,
 	setPinMappings,
-	getProfileOptions,
-	setProfileOptions,
 	getKeyMappings,
 	setKeyMappings,
 	getAddonsOptions,
